+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.4.1 - 2022-02-18
+
+#### features
+
+* draw cutting lines with scissor symbol
+
+#### bug fixes
+
+* fix empty line after addresses when using combined address
+
 ### 0.4.0 - 2022-02-18
 
 * initial release